move validator typeclasses to validation syntax
