--- application/src/main/scala/com/azavea/franklin/extensions/validation/ItemValidator.scala
@@ -1,37 +1,0 @@
-package com.azavea.franklin.extensions.validation
-
-import cats.syntax.semigroup._
-import com.azavea.stac4s.StacItem
-import com.azavea.stac4s.extensions.ItemExtension
-import com.azavea.stac4s.extensions.eo.EOItemExtension
-import com.azavea.stac4s.extensions.label.LabelItemExtension
-import com.azavea.stac4s.extensions.layer.LayerItemExtension
-import com.azavea.stac4s.syntax._
-import eu.timepit.refined.auto._
-import eu.timepit.refined.types.string.NonEmptyString
-
-trait ItemValidator[T] {
-  def validate(item: StacItem): StacItem
-}
-
-object ItemValidator {
-
-  def apply[T](implicit ev: ItemValidator[T]) = ev
-
-  def instance[T: ItemExtension](name: NonEmptyString) = new ItemValidator[T] {
-
-    def validate(item: StacItem) = {
-      val extensionResult     = item.getExtensionFields[T]
-      val validationExtension = ValidationExtension.fromResult(extensionResult, name)
-      val existingValidation  = item.getExtensionFields[ValidationExtension]
-      item.addExtensionFields(existingValidation map { _ |+| validationExtension } getOrElse {
-        validationExtension
-      })
-    }
-  }
-
-  implicit val labelItemValidator: ItemValidator[LabelItemExtension] = instance("label")
-  implicit val layerItemValidator: ItemValidator[LayerItemExtension] = instance("layer")
-  implicit val eoItemValidator: ItemValidator[EOItemExtension]       = instance("eo")
-
-}
